backend/requirements.txt
# Package configuration
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
pillow==10.1.0
pandas==2.1.3
pytest==7.4.3
httpx==0.25.1
python-dateutil==2.8.2
bcrypt==4.0.1
pyotp==2.8.0
qrcode[pil]==7.4.2
jinja2==3.1.4
yfinance==0.2.65